backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1
boto3==1.34.34
psycopg2-binary==2.9.9
google-auth==2.27.0
requests==2.31.0
google-cloud-monitoring==2.29.1
Pillow==10.2.0
sse-starlette==2.1.0
asgiref==3.7.2
pgvector==0.4.2

# ML Dependencies for RAG (ONNX)
# Pinning versions for stability (CPU inference)
numpy<2.0
onnx==1.16.1
onnxscript>=0.1.0
optimum==1.19.1
onnxruntime==1.18.0
sentence-transformers==2.5.1
